requirements-apt.txt
python3
python3-dbus
python3-gnupg
python3-pyclamd
